% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{add_habitat}
\alias{add_habitat}
\title{Add habitat}
\usage{
add_habitat(base, watershed, amount, years)
}
\arguments{
\item{watershed}{watershed name or number 1:31}

\item{amount}{square meters}

\item{years}{vector of characters 1980-1999}
}
\description{
Add habitat
}
