% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{create_scenario}
\alias{create_scenario}
\title{Create scenario}
\usage{
create_scenario(scenario_df, species = c("fr", "wr", "sr"))
}
\arguments{
\item{scenario_df}{a dataframe containing scenario information, see details below}

\item{species}{either "fr", "wr", "sr" for fall run, winter run, or spring run respectively}
}
\description{
Create scenario
}
\details{
The \code{scenario_df} is a dataframe with each row representing a scenario action.
The dataframe must contain the following columns:
\itemize{
  \item "watershed" - The name or index for a watershed, ex. "Upper Sacramento River" or 1
  \item "action" - The action taken represented by the code 1 - 5, ex: 3
  \item "start_year"- The simulation year the action begins
  \item "end_year" - The simulation year the action ends
  \item "units_of_effort" - number of action units taken, ex: .5 or 1
}
}
\section{Actions and Units of Effort}{

\itemize{
  \item Action 1: Do nothing
  \item Action 2: Add 1 acre of spawning habitat
  \item Action 3: Add 2 acres of inchannel rearing habitat
  \item Action 4: Add 2 acres of floodplain rearing habitat
  \item Action 5: Increase rearing survival by 5%

}
}

\examples{
scenario_df <- data.frame(watershed = c("Upper Sacramento River", "Upper Sacramento River",
                        "American River", "Feather River", "Lower-mid Sacramento River",
                        "Battle Creek", "Butte Creek", "Deer Creek", "Stanislaus River"),
          action = c(3, 3, 3, 3, 3, 3, 3, 3, 3),
          start_year = c(1980, 1990, 1980, 1980, 1980, 1990, 1990, 1990, 1990),
          end_year = c(1989, 1999, 1989, 1989, 1989, 1999, 1999, 1999, 1999),
          units_of_effort = c(2, 1, 1, 1, 1, 1, 1, 1, 1))

create_scenario(scenario_df, species = "fr")
}
