% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decay_rate}
\alias{decay_rate}
\alias{spawn_decay_rate}
\alias{rear_decay_rate}
\title{Decay Rates}
\format{
An object of class \code{numeric} of length 31.

An object of class \code{numeric} of length 31.
}
\usage{
spawn_decay_rate

rear_decay_rate
}
\description{
Decay Rates
}
\keyword{datasets}
