% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{get_action_matrices}
\alias{get_action_matrices}
\title{Get Action Matrices}
\usage{
get_action_matrices(scenario_df)
}
\arguments{
\item{scenario_df}{scenario dataframe}
}
\description{
Converts scenario dataframe into a list of matrices [31 watersheds, years]
   for each action type. The matrix values are the number of units of effort applied
   for that action type within a watershed for that year.
}
