% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{increase_survial_scalar}
\alias{increase_survial_scalar}
\title{Increase Survival}
\usage{
increase_survial_scalar(base, watershed, amount = 1.05, years)
}
\arguments{
\item{watershed}{watershed name or number 1:31}

\item{amount}{1 unit of effort is a 5\% increase, default 1 unit is 1.05}

\item{years}{vector of characters 1980-1999}
}
\description{
Increase Survival
}
