% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{modify_spawning_habitat}
\alias{modify_spawning_habitat}
\title{Modify Spawning Habitat
TODO can decay be passed in as argument? for sensitivity analysis}
\usage{
modify_spawning_habitat(
  habitat,
  action_units,
  amount,
  theoretical_max,
  stochastic,
  decay_multipliers
)
}
\description{
Modify Spawning Habitat
TODO can decay be passed in as argument? for sensitivity analysis
}
