% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{SIT Scenarios}
\format{
An object of class \code{list} of length 14.
}
\usage{
DSMscenario::scenarios$NO_ACTION
DSMscenario::scenarios$ONE
DSMscenario::scenarios$TWO
DSMscenario::scenarios$THREE
DSMscenario::scenarios$FOUR
DSMscenario::scenarios$FIVE
DSMscenario::scenarios$SIX
DSMscenario::scenarios$SEVEN
DSMscenario::scenarios$EIGHT
DSMscenario::scenarios$NINE
DSMscenario::scenarios$TEN
DSMscenario::scenarios$ELEVEN
DSMscenario::scenarios$TWELVE
DSMscenario::scenarios$THIRTEEN
}
\description{
CVPIA SIT's scenario dataframes for scenarios 1-7 and no action
}
\examples{
fall_run_seeds <- fallRunDSM::fall_run_model(mode = "seed")
fallRunDSM::fall_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = fall_run_seeds)
}
\keyword{datasets}
