% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_groups}
\alias{watershed_groups}
\title{SIT Watershed Groupings}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
watershed_groups
}
\description{
CVPIA SIT's watershed groupings
}
\keyword{datasets}
