% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{load_scenario}
\alias{load_scenario}
\title{Load Scenario}
\usage{
load_scenario(
  scenario,
  habitat_inputs,
  species = c("fr", "wr", "sr", "st", "lfr"),
  spawn_decay_rate = DSMscenario::spawn_decay_rate,
  rear_decay_rate = DSMscenario::rear_decay_rate,
  stochastic = TRUE
)
}
\arguments{
\item{habitat_inputs}{a list with spawning habitat, inchannel fry and juvenile rearing habitat,
floodplain rearing habitat, and weeks flooded matrices from the lifecycle model
"params" data object. See the example for more details.}

\item{species}{provide \code{"fr"}, \code{"lfr"}, \code{"wr"}, \code{"sr"}, or \code{"st"} for fall run, late-fall run,
winter run, spring run, or steelhead respectively to designate which \code{params} data object
to be modified. For example, supply \code{"fr"} if running the \code{fallRunDSM::fall_run_model}.}

\item{spawn_decay_rate}{length 31 vector of 1 - spawning decay rate estimates}

\item{rear_decay_rate}{length 31 vector of 1 - rearing decay rate estimates}

\item{stochastic}{boolean, TRUE for creating scenarios with stochasticity}

\item{scenario_df}{a dataframe containing scenario information, see details below}
}
\description{
Modify baseline model inputs based on a set of actions
}
\details{
The \code{scenario_df} is a dataframe with each row representing a scenario action.
The dataframe must contain the following columns:
\itemize{
  \item \strong{watershed} - The name or index for a watershed, ex. "Upper Sacramento River" or 1
  \item \strong{action} - The action taken represented by the code 1 - 5, ex: 3 to represent adding inchannel rearing
  \item \strong{start_year} - The simulation year the action begins
  \item \strong{end_year} - The simulation year the action ends
  \item \strong{units_of_effort} - number of action units taken, ex: .5 or 1
}
}
\section{Actions and Units of Effort}{

\itemize{
  \item \strong{1} - Do nothing
  \item \strong{2} - Add 1 acre of spawning habitat
  \item \strong{3} - Add 2 acres of inchannel rearing habitat
  \item \strong{4} - Add 3 acres of floodplain rearing habitat
  \item \strong{5} - Increase rearing survival by 5\%

}
}

\examples{
habitats <- list(
  spawning_habitat = fallRunDSM::params$spawning_habitat,
  inchannel_habitat_fry = fallRunDSM::params$inchannel_habitat_fry,
  inchannel_habitat_juvenile = fallRunDSM::params$inchannel_habitat_juvenile,
  floodplain_habitat = fallRunDSM::params$floodplain_habitat,
  weeks_flooded = fallRunDSM::params$weeks_flooded
)

scenario <- load_scenario(scenario = DSMscenario::scenarios$ONE,
                          species = DSMscenario::species$FALL_RUN,
                          habitat_inputs = habitats)
}
