% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{SIT Scenarios}
\format{
An object of class \code{list} of length 14.
}
\usage{
scenarios
}
\description{
CVPIA SIT's scenario dataframes for scenarios 1-13 and no action
}
\details{
Subset DSMscenario::scenarios to access scenario data for a predefined SIT strategy.
View \href{https://flowwest.shinyapps.io/SIT-strategies/}{SIT-strategies}.
}
\examples{
# Select a scenario
scenario_one <- DSMscenario::scenarios$ONE

# Use that scenario in the fall run model
fall_run_seeds <- fallRunDSM::fall_run_model(mode = "seed")
fallRunDSM::fall_run_model(scenario = scenario_one,
                           mode = "simulate",
                           seeds = fall_run_seeds)
}
\keyword{datasets}
